+++ owl-back/src/main/java/com/project/owlback/codereview/dto/CodeReviewCommentReqDto.java
@@ -10,14 +10,14 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CodeReviewCommentReqDto {
-    private Integer codeHistoryId;
-    private Integer parent;
+    private Long codeHistoryId;
+    private Long parent;
     private Integer depth;
     private String contents;
     private Integer startLine;
     private Integer endLine;
-    private Integer codeReviewId;
+    private Long codeReviewId;
     private Integer versionNum;
-    private Integer codeCommentId;
+    private Long codeCommentId;
     private boolean like;
 }
